There's the app!
I tried to clean this up a little bit...so many regrets from 2012. There
were also a couple of things in the code that caused me to pull the
original repository in the first place (aka, they shouldn't be exposed),
which unfortunately led me to breaking one or two functions. That said,
the main functionality of the program remains intact.